requirements.txt
semantic-kernel
openai>=1.52.0
psycopg2
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.38.0
